requirements.txt
openai-whisper==20231117
pandas==2.2.3
pyyaml==6.0.1
pydub==0.25.1
openpyxl==3.1.5
watchdog==3.0.0
openai>=1.3.5 